move engine debugging functions to separate file and struct
